requirements.txt
hydra-core==1.3.2
matplotlib==3.7.0
numpy==1.24.2
scikit-learn==1.2.1
wandb==0.14.0